Fix: All Windows Pop-up scaling on hidpi monitors
After right clicking and opening the "All Windows" tab, the pop-up
will appear cut off on some hidpi monitors.

It appears that the logic for rendering doesn't have any cap
for max height ( or left) so that it goes negative coords,
which then the more you add, the more negative the rendering goes.