fix: during initialization, set animation out time to 0
If manual hide is enabled, then we do not want to animate the dock at
all.  However, the _animateOut function performs some side-effects on
completion, so we need to run it.  Therefore just set the animation out
timeout to 0 for all modes.

To distinguish between normal behavior and initialization, introduce a
boolean init variable to _updateDashMode and _updateDashVisibility that
by default is set to false and thus performs the old behavior.  In
_initialize, however, we call _updateDashMode with this parameter to
true and thus eliminate the animation time.

The outcome of this change is that unlocking the screen or reloading the
extension will not animate the dock if it is set to manual hide.

To properly re-enable the dock, call the untrack and track Dock methods.