fix: during initialization, set animation out time to 0
If `manualhide` is enabled, then we do not want to animate the dock at
all.  However, the `_animateOut` function performs some side-effects on
completion, so we need to run it.  Therefore just set the animation out
timeout to 0 for all modes.

To distinguish between initialization and normal behavior, we set a
`_initializing` attribute to `true` in `_init`, and then set it to
`false` at the end of `_initialize()`.

The outcome of this change is that unlocking the screen or reloading the
extension will not animate the dock if it is set to `manualhide` while
keeping support for the other dock modes.